requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

psycopg2-binary==2.9.9
Flask==3.0.3
APScheduler==3.10.4

selenium==4.24.0
python-dateutil==2.9.0.post0
fpdf==1.7.2
Pillow==10.1.0

# Use modern OpenAI client (v1+). Pin to a compatible release range.
openai>=1.0.0,<3.0.0

